rough outline for of projectile scene
test scene for everyone to join scripts

+++ Zombie Project/Assets/ShaneTesting/Scripts/Barricade.cs
@@ -6,7 +6,7 @@ public class Barricade : MonoBehaviour
  {
 
 
-    public int barricadeHealth;
+    public float barricadeHealth;
     public ZombieAI zombieAIScript;
